backend/requirements.txt
Flask
flask-cors
numpy
google-cloud-aiplatform
google-cloud-firestore
